refactor(st-09064): modularize middleware presets
