+++ packages/core/src/langgraph/middleware/preset-collection.ts
@@ -1,0 +1,9 @@
+import { development } from './preset-development.js';
+import { production } from './preset-production.js';
+import { testing } from './preset-testing.js';
+
+export const presets = {
+  production,
+  development,
+  testing,
+};
